+++ Rdmp.Core.Tests/DataLoad/Engine/Unit/CommandLineHelperTests.cs
@@ -5,7 +5,7 @@
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
 using System;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.IO;
 using FAnsi.Discovery;
 using FAnsi.Implementation;