+++ Rdmp.Core/DataLoad/Modules/Mutilators/StagingBackfillMutilator.cs
@@ -6,7 +6,7 @@
 
 using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Linq;
 using FAnsi.Discovery;
 using Rdmp.Core.Curation.Data;